+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Changed
 - Метод [RepositoryInterface::getModelClass](src/Repository/RepositoryInterface.php) метод теперь должен быть публичным
-- [RepoHelper.php](src/Helpers/RepoHelper.php) в большинстве методов теперь может принимать третьим параметром не только 
+- [RepoHelper.php](src/Helpers/ReflectionHelper.php) в большинстве методов теперь может принимать третьим параметром не только 
 массив, но и скалярные значения
 - Изменен порядок и тип передаваемых аргументов в  [ModelManager::setModelConfig](src/Manager/ModelManager.php). Теперь
 в качестве первого аргумента идет репозиторий, а вторым аргументом идет модель. Если не указывать модель, то ее класс будет